% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesb.R
\docType{data}
\name{cetesb}
\alias{cetesb}
\title{Location some air quality stations CETESB}
\format{A spatial data-frame with locations of CETESB stations, class sf.
Extracted from the report of air quality 2018
\describe{
  \item{red}{type of station}
  \item{UGRHI}{integer}
  \item{VOCACIONAL}{Character}
  \item{UTM_SIRGAS2000}{character}
  \item{fuso}{integer}
  \item{EPSG}{integer}
  \item{ENDERECO}{character}
  \item{OBSERVACOES}{Observations}
  \item{id}{integer for Rpollution or koffing}
  \item{X}{Numeric}
  \item{Y}{Numeric}
  \item{stationname}{Character}
  \item{initial_date}Character
  \item{BEN}{Character}
  \item{CO}{Character}
  \item{ERT}{Character}
  \item{MP10}{Character}
  \item{MP2.5}{Character}
  \item{NO}{Character}
  \item{NO2}{Character}
  \item{NOx}{Character}
  \item{O3}{Character}
  \item{SO2}{Character}
  \item{TOL}{Character}
  \item{TEMP}{Character}
  \item{UR}{Character}
  \item{VV}{Character}
  \item{address}{Character}
  \item{UTM_SIRGAS_2000}{Character}
  \item{geometry}{Character}
}}
\source{
\url{https://cetesb.sp.gov.br/ar/}
}
\usage{
data(cetesb)
}
\description{
Location some air quality stations CETESB
}
\keyword{datasets}
