% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covnox.R
\name{covnox}
\alias{covnox}
\title{Divide data.frames}
\usage{
covnox(cov, nox, geo)
}
\arguments{
\item{geo}{'sf'}

\item{X}{Character; path to .rds}

\item{name}{Character; Name of new column.}
}
\value{
Return data.frame or 'sf'
}
\description{
Return data.frame sf
}
\examples{
\dontrun{
}
}
