% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curitiba.R
\docType{data}
\name{curitiba}
\alias{curitiba}
\title{Location some air quality stations CETESB}
\format{
A spatial data-frame with locations of CETESB stations, class sf.
\describe{
  \item{Station}{Character}
  data(curitiba)
}
}
\source{
\url{http://www.iat.pr.gov.br/Pagina/Monitoramento-da-Qualidade-do-Ar}
}
\usage{
data(curitiba)
}
\description{
A spatial data-frame with locations of CETESB stations, class sf.
}
\keyword{datasets}
