% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_wrf.R
\name{project_wrf}
\alias{project_wrf}
\title{project your wrfout files to any EPSG projection}
\usage{
project_wrf(file = file.choose(), map_proj = "lambert", truelat1,
  truelat2, stand_lon, epsg_target = "+init=epsg:32719",
  map_proj_char_target = "UTM", backup_path = getwd(),
  verbose = TRUE)
}
\arguments{
\item{file}{name of file interactively (default) or specified}

\item{map_proj}{"lambert" or "mercator"}

\item{truelat1}{return a raster instead of an array}

\item{truelat2}{crs to use if as_raster is TRUE}

\item{stand_lon}{level for rasters from a 4D variable}

\item{epsg_target}{level for rasters from a 4D variable}

\item{map_proj_char_target}{level for rasters from a 4D variable}

\item{backup_path}{level for rasters from a 4D variable}

\item{verbose}{level for rasters from a 4D variable}
}
\description{
Read a wrfout file, project its coordinates and put them back
}
\examples{
\dontrun{
# dont run
project_wrf(file = "wrfout_d01_2016-12-31_00:00:00",
            truelat1 = -30,
            truelat2 = -60,
            stand_lo n =-71.626119)
}
}
