% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmeans_df.R
\name{rmeans_df}
\alias{rmeans_df}
\title{Calculates rowMeans reading .rds}
\usage{
rmeans_df(X, name, geo)
}
\arguments{
\item{X}{Character; path to .rds}

\item{name}{Character; Name of new column.}

\item{geo}{'sf'}
}
\value{
Return data.frame or 'sf'
}
\description{
Return data.frame sf
}
\examples{
\dontrun{
#do not run
}
}
