% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_wrf.R
\name{st_bbox_wrf}
\alias{st_bbox_wrf}
\title{st_bbox_wrf from wrf files}
\usage{
st_bbox_wrf(filewrf, type = "geo", crs = 4326)
}
\arguments{
\item{filewrf}{wrf file}

\item{type}{Type of wrf file: "input", "geo", "out", "chemi", "bdy"}

\item{crs}{epsg code number (see http://spatialreference.org/ref/epsg/)}
}
\description{
\code{\link{st_bbox_wrf}} from wrf files using the corner coordinates
\code{\link{st_bbox_wrf_input}} is for wrf inputs
\code{\link{st_bbox_wrf_chemi}} is for wrf chemi inputs #TODO CHECK
\code{\link{st_bbox_wrf_bdy}} is for wrf boundaries #TODO CHECK
\code{\link{st_bbox_wrf_out}} is for wrf outs #TODO CHECK
\code{\link{st_bbox_wrf_geo}} is for wrf outs #TODO CHECK
}
\examples{
{
# Do not run
wrf <- paste(system.file("extdata", package = "eixport"),
                         "/wrfinput_d02", sep="")
gwrf  <- st_bbox_wrf_input(wrf)
plot(gwrf, axes = TRUE)
}
}
