% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_wrf.R
\name{st_bbox_wrf_geo}
\alias{st_bbox_wrf_geo}
\title{st_bbox_wrf_geo}
\usage{
st_bbox_wrf_geo(filewrf, crs = 4326)
}
\arguments{
\item{filewrf}{wrf file}

\item{crs}{epsg code number (see http://spatialreference.org/ref/epsg/)}
}
\description{
st_bbox_wrf_geo
}
