% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_df.R
\name{sum_df}
\alias{sum_df}
\title{Read .rds and sum data.frames returning data.frame or sf}
\usage{
sum_df(X, geo)
}
\arguments{
\item{X}{Character; path to .rds}

\item{geo}{'sf'}
}
\value{
Return data.frame or 'sf'
}
\description{
Return data.framee or sf
}
\examples{
\dontrun{
}
}
