% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{Compare observations and simulations and apply wilcoxon R test}
\usage{
verify(
  dfobs,
  dfmod,
  x,
  Station = "Station",
  time = "LT",
  time_spinup,
  conf.level.p = 0.95,
  conf.level.w = 0.95
)
}
\arguments{
\item{dfobs}{data.frame in long format with columns must be Station, x and time}

\item{dfmod}{data.frame in long format with columns must be Station, x and time}

\item{x}{Character; name of the column with the parameter}

\item{Station}{Character; name of the column with the stations}

\item{time}{Character; name of the time column with class POSIXct.}

\item{time_spinup}{Character; time to filter}

\item{conf.level.p}{Pearson confidence level for the returned confidence interval.
Currently only used for the Pearson product moment correlation coefficient
if there are at least 4 complete pairs of observations. default = 0.95}

\item{conf.level.w}{Wilcox confidence level of the interval. default = 0.95}
}
\value{
Return data.framee or list of raster and df
}
\description{
Return data.frame with Correlation, MeanBias, SD, RSME and wilcoxon p.value
}
\examples{
\dontrun{
# do not run
# Columns must be Station, x and time
verify(dfobs = df_co[df_co$data == "CETESB", 1:3],
       dfmod = df_co[df_co$data == "2016", 1:3],
       x = "co",
       time_spinup = ISOdatetime(2016,4,8,12,0,0))
}
}
