% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtractor_raster.R
\name{xtractor_raster}
\alias{xtractor_raster}
\title{Extract several variables from RasterBricks}
\usage{
xtractor_raster(
  br,
  points,
  times,
  crs_points = 4326,
  tz = "America/Sao_Paulo",
  verbose = TRUE,
  allow.cartesian = TRUE
)
}
\arguments{
\item{br}{RasterBrick}

\item{points}{data.frame, matrix, SpatialPointsDataFrame or sf 'POINTS',
with coordinates East-weast (lat), north-south (long) and \strong{Station}.}

\item{times}{Character; UTC time (one time for each layer)}

\item{crs_points}{Integer, crs points.}

\item{tz}{Character; Time zone}

\item{verbose}{logical}

\item{allow.cartesian}{logical (from data.table):
prevents joins that would result in more than nrow(x)+nrow(i) rows.
This is usually caused by duplicate values in i's join columns,
each of which join to the same group in 'x' over and over again:
a misspecified join. Usually this was not intended and the join
needs to be changed. The word 'cartesian' is used loosely in
this context. The traditional cartesian join is (deliberately)
difficult to achieve in data.table: where every row in i joins to every
row in x (a nrow(x)*nrow(i) row result). 'cartesian' is just meant
in a 'large multiplicative' sense, so FALSE does not always prevent a
traditional cartesian join.}
}
\value{
Return data.framee or list of raster and df
}
\description{
Return data.framee or list of raster and df from points. When is data.frame
contains cell value, interpolation for lat long based on neighour cells,
}
\note{
Based on NCL scrip xtractor from DCA/IAG/USP
}
\examples{
\dontrun{
# do not run
# br <- brick("wrfout")
data(cetesb)
df <- xtractor_raster(br = br,  points = cetesb)
}
}
