% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetesb.R
\docType{data}
\name{cetesb}
\alias{cetesb}
\title{Location some air quality stations CETESB}
\format{A data frame with 288 rows and 12 variables:
\describe{
  \item{Station}{Station}
  \item{geometry}{geometry}
}}
\source{
\url{http://cetesb.sp.gov.br/ar/publicacoes-relatorios/}
}
\usage{
data(cetesb)
}
\description{
Location some air quality stations CETESB
}
\keyword{datasets}
