% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covnox.R
\name{covnox}
\alias{covnox}
\title{Divide data.frames}
\usage{
covnox(cov, nox, geo)
}
\arguments{
\item{cov}{Data-frame of numerics}

\item{nox}{Data-frame of numerics}

\item{geo}{'sf'}
}
\value{
Return data.frame or 'sf'
}
\description{
Return data.frame sf
}
\examples{
\dontrun{
#do not run
}
}
