% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtractor.R
\name{xtractor}
\alias{xtractor}
\title{Extract several variables from WRF outputs}
\usage{
xtractor(atmos, vars, level = 1, points, stations = "CC",
  crs_points = 4326, model = "WRF", tz = "America/Sao_Paulo",
  return_list = FALSE, verbose = TRUE)
}
\arguments{
\item{atmos}{Character; path to output of model (wrfout)}

\item{vars}{Character; Variable with has 3 or 4 dimensions.}

\item{level}{Integer; Which level}

\item{points}{data.frame, matrix, SpatialPointsDataFrame or sf 'POINTS',
with coordinates East-weast (lat), north-south (long) and 'Station'.}

\item{stations}{Character; names of stations for each point.}

\item{crs_points}{Integer, crs points.}

\item{model}{Character; Currently, only "WRF"}

\item{tz}{Character; Time zone}

\item{return_list}{Logical; If TRUE, return a list with raster and data.frames
if FALSE, only data.frame.}

\item{verbose}{logical}
}
\value{
Return data.framee or list of raster and df
}
\description{
Return data.framee or list of raster and df from points.
}
\note{
Based on NCL scrip xtractor from DCA/IAG/USP
}
\examples{
\dontrun{
data(cetesb)
#use your wrfout
#wrf <- "~/Documentos/wrfo/wrfoA.nc"
t2 = c("T2", "o3", "no")
df <- xtractor(atmos = wrf, vars = t2, points = cetesb,
stations = cetesb$Station[1:3])
r <- xtractor(atmos = wrf, vars = t2, points = cetesb return_list = T)
}
}
