% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtractor_raster.R
\name{xtractor_raster}
\alias{xtractor_raster}
\title{Extract several variables from RasterBricks}
\usage{
xtractor_raster(
  br,
  var,
  points,
  times,
  crs_points = 4326,
  tz = "America/Sao_Paulo",
  verbose = TRUE
)
}
\arguments{
\item{br}{RasterBrick}

\item{var}{Character; 1 word}

\item{points}{data.frame, matrix, SpatialPointsDataFrame or sf 'POINTS',
with coordinates East-weast (lat), north-south (long) and \strong{Station}.}

\item{times}{Character; UTC time (one time for each layer)}

\item{crs_points}{Integer, crs points.}

\item{tz}{Character; Time zone}

\item{verbose}{logical}
}
\value{
Return data.framee or list of raster and df
}
\description{
Return data.framee or list of raster and df from points. When is data.frame
contains cell value, interpolation for lat long based on neighour cells,
}
\note{
Based on NCL scrip xtractor from DCA/IAG/USP
}
\examples{
\dontrun{
# do not run
# br <- brick("wrfout")
data(cetesb)
df <- xtractor_raster(br = br, vars = "T2", points = cetesb)
}
}
